# swagger.yaml
openapi: 3.0.0
info:
  title: Express API
  version: 1.0.0
  description: API documentation for the Express application
servers:
  - url: http://localhost:3000
paths:
  /accounts:
    post:
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  email:
                    type: string
        '400':
          description: Invalid request
        '409':
          description: Account with this email already exists
        '500':
          description: Server error
  /sessions:
    post:
      summary: Create a new session (login)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  accountId:
                    type: integer
        '400':
          description: Invalid request
        '401':
          description: Invalid password or account not found
        '500':
          description: Server error
    delete:
      summary: Delete a session (logout)
      responses:
        '204':
          description: Session deleted
        '401':
          description: Unauthorized
        '500':
          description: Server error
  /expenses:
    post:
      summary: Create a new expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                amount:
                  type: number
      responses:
        '201':
          description: Expense created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  amount:
                    type: number
                  userId:
                    type: integer
        '400':
          description: Invalid request
        '500':
          description: Server error
    get:
      summary: Retrieve expenses
      responses:
        '200':
          description: Expenses retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    amount:
                      type: number
                    userId:
                      type: integer
        '500':
          description: Server error
  /incomes:
    post:
      summary: Create a new income
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                amount:
                  type: number
      responses:
        '201':
          description: Income created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  amount:
                    type: number
                  userId:
                    type: integer
        '400':
          description: Invalid request
        '500':
          description: Server error
    get:
      summary: Retrieve incomes
      responses:
        '200':
          description: Incomes retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    amount:
                      type: number
                    userId:
                      type: integer
        '500':
          description: Server error
  /expenses/{id}:
    put:
      summary: Update an expense
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                amount:
                  type: number
      responses:
        '200':
          description: Expense updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  amount:
                    type: number
        '400':
          description: Invalid request
        '404':
          description: Expense not found
        '500':
          description: Server error
    delete:
      summary: Delete an expense
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Expense deleted
        '404':
          description: Expense not found
        '500':
          description: Server error
  /incomes/{id}:
    put:
      summary: Update an income
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                amount:
                  type: number
      responses:
        '200':
          description: Income updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  amount:
                    type: number
        '400':
          description: Invalid request
        '404':
          description: Income not found
        '500':
          description: Server error
    delete:
      summary: Delete an income
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Income deleted
        '404':
          description: Income not found
        '500':
          description: Server error
  /support_tickets:
    post:
      summary: Create a new support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subject:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Support ticket created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  subject:
                    type: string
                  description:
                    type: string
                  status:
                    type: string
                  created_at:
                    type: string
        '400':
          description: Invalid request
        '500':
          description: Server error
    get:
      summary: Retrieve support tickets for the authenticated user
      responses:
        '200':
          description: Support tickets retrieved